update: move minigame mapper -> shop mapper
