+++ src/main/kotlin/gogo/gogoshop/domain/shop/root/application/ShopMapper.kt
@@ -3,9 +3,12 @@ package gogo.gogoshop.domain.shop.root.application
 
 
 import gogo.gogoshop.domain.cointoss.application.dto.CoinTossResDto
+import gogo.gogoshop.domain.cointoss.persistence.CoinToss
 import gogo.gogoshop.domain.plinko.application.dto.PlinkoResDto
+import gogo.gogoshop.domain.plinko.persistence.Plinko
 import gogo.gogoshop.domain.shop.root.application.dto.ShopTicketStatusResDto
 import gogo.gogoshop.domain.yavarwee.application.dto.YavarweeResDto
+import gogo.gogoshop.domain.yavarwee.persistence.Yavarwee
 import org.springframework.stereotype.Component
 
 @Component
@@ -23,5 +26,32 @@ class ShopMapper {
             yavarwee = yavarweeDto,
             plinko = plinkoDto
         )
+
+    fun map(coinToss: CoinToss?) =
+        coinToss?.let {
+            CoinTossResDto(
+                coinTossId = it.coinTossId,
+                ticketPrice = it.ticketPrice,
+                ticketQuantity = it.ticketQauntity
+            )
+        }
+
+    fun map(plinko: Plinko?) =
+        plinko?.let {
+            PlinkoResDto(
+                plinkoId = it.plinkoId,
+                ticketPrice = it.ticketPrice,
+                ticketQuantity = it.ticketQauntity
+            )
+        }
+
+    fun map(yavarwee: Yavarwee?) =
+        yavarwee?.let {
+            YavarweeResDto(
+                yavarweeId = it.yavarweeId,
+                ticketPrice = it.ticketPrice,
+                ticketQuantity = it.ticketQauntity
+            )
+        }
 
 }
